renamed jax_toy_text to tabular
