+++ gymnasium/envs/__init__.py
@@ -161,7 +161,7 @@ register(
 
 register(
     id="Jax-BlackJack-v0",
-    entry_point="gymnasium.envs.jax_toy_text.blackjack:BlackJackJaxEnv",
+    entry_point="gymnasium.envs.tabular.blackjack:BlackJackJaxEnv",
     kwargs={"sutton_and_barto": True, "natural": False},
 )
