--- gymnasium/envs/jax_toy_text/__init__.py
@@ -1,1 +1,0 @@
-from gymnasium.envs.jax_toy_text.blackjack import BlackJackJaxEnv
