+++ gymnasium/envs/tabular/__init__.py
@@ -1,0 +1,1 @@
+from gymnasium.envs.tabular.blackjack import BlackJackJaxEnv
